global_chem_extensions_package/requirements.txt
plotly
global-chem
kaleido
rdkit-pypi
bokeh==1.4.0
